+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/chat/resp/ChatMessageResponse.java
@@ -1,7 +1,7 @@
 package cc.xmist.mistchat.server.chat.resp;
 
-import cc.xmist.mistchat.server.common.enums.RoomType;
 import cc.xmist.mistchat.server.common.enums.MessageType;
+import cc.xmist.mistchat.server.common.enums.RoomType;
 import lombok.Builder;
 import lombok.Data;
 
@@ -9,15 +9,15 @@ import java.time.LocalDateTime;
 
 @Data
 public class ChatMessageResponse {
-    private Long uid;
+    private Integer uid;
     private RoomType type;
     private Message message;
 
     @Data
     @Builder
     public static class Message {
-        private Long id;
-        private Long roomId;
+        private Integer id;
+        private Integer roomId;
         private RoomType roomType;
         private LocalDateTime sendTime;
         private MessageType type;
