+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/socketio/EventEmitter.java
@@ -12,17 +12,17 @@ public class EventEmitter {
 
     private final ClientPool clientPool;
 
-    private void emit(Long uid, String name, Object data) {
+    private void emit(Integer uid, String name, Object data) {
         if (clientPool.isOnline(uid)) {
             clientPool.get(uid).sendEvent(name, data);
         }
     }
 
-    public void emit(Long uid, BaseEvent event) {
+    public void emit(Integer uid, BaseEvent event) {
         emit(uid, event.getName(), event.getData());
     }
 
-    public void emits(List<Long> uids, BaseEvent event) {
+    public void emits(List<Integer> uids, BaseEvent event) {
         uids.forEach(uid -> {
             emit(uid, event.getName(), event.getData());
         });
