+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/user/controller/UserController.java
@@ -27,7 +27,7 @@ public class UserController {
     @Operation(summary = "获取当前用户信息")
     @GetMapping("/info")
     public R<UserResponse> getUserInfo() {
-        Long uid = RequestContext.getUid();
+        Integer uid = RequestContext.getUid();
         UserResponse user = userService.getUserInfo(uid);
         return R.ok(user);
     }
@@ -35,14 +35,14 @@ public class UserController {
     @Operation(summary = "其他用户信息")
     @PostMapping("/another/info/list")
     public R<List<UserResponse>> getUserInfoBatched(@RequestBody cc.xmist.mistchat.server.user.model.req.UidListRequest req) {
-        Long uid = RequestContext.getUid();
+        Integer uid = RequestContext.getUid();
         return R.ok(userService.getBatchedUserInfo(uid, req.getUidList()));
     }
 
     @Operation(summary = "修改用户名称")
     @PostMapping("/name/modify")
     public R<Void> modifyName(@RequestBody @Valid cc.xmist.mistchat.server.user.model.req.ModifyNameRequest modifyNameReq) {
-        Long uid = RequestContext.getUid();
+        Integer uid = RequestContext.getUid();
         userService.modifyName(uid, modifyNameReq.getName());
         return R.ok();
     }
